core:admin: add s3 backup support to sailor
this includes only the functionality definition, the integration is yet to be done. This contains backing up raw .sail files which is the main database of the sailor and the current app state of each project configured under sailor.

The app state is then used by the application as fallback incase it couldn't reach the sailor service