sdk: use atomics and snapshots instead of RWMutex
while using RWMutex there might be scenarios where the developer might not Release the Reader Lock which might lead to Deadlock, where Writer will never be able to Write to the configs.

Here each goroutine will get their own snapshotted version of the config, and can be passed around during the lifespan of the thread, this will make sure we have consistent reads and developers need not know the implementation behind the scenes.