feat(cli): add sync and init commands for interacting with sailor file (#13)
a sailor file defines, maintains the resources used by a project. The resources are defined inside `sailor.json` and the `sync` command is used to download the latest version of the mentioned resource inside the file provided in the config. This ensures that everyone develops over the latest config for that environment.