feat(sailor): re-architect the authorization flow
This is a sailor level re-write where we have changed the security level from project level to cluster-level. The idea here is to begin simple at start and then build features that are really necessary in a extensible way. This will be a no brainer for many users to set a single access and secret key per environment/cluster and move forward with their integrations.

This commit contains the following high-level features:
- Sailor now operates fully on token and the token now has access and secret key
- Clients can exchange token for key pair when secret decryption is required
- Projects now can be simply created without keypars being provided for each one
- Internals sail related operations are now being migrated into `Sail` interface slowly for better testability