feature(core): add support for adding s3 settings and deployment
this standardizes the s3 backup feature which can be enabled for each
resources by a setting `deploy.s3=true`. This lets the app owner deside
to deploy it on s3 and use it as a fallback incase admission hook fails
or volume mount fails for some reason. This fallback will serve as a
last resort even when sailor core is not available and apps can still
fetch the latest config from s3.